chore: rework contest entity's model and redux store
