+++ src/entities/contest/index.ts
@@ -1,0 +1,6 @@
+import { contestActions, contestReducer } from './model/slice'
+import { Contest, ContestState } from './model/types'
+
+export type { Contest, ContestState }
+
+export { contestActions, contestReducer }
