+++ src/entities/contest/model/selectors/index.ts
@@ -1,0 +1,5 @@
+export const selectContestData = (state: RootState) => state.contest.data
+
+export const selectContestLoading = (state: RootState) => state.contest.loading
+
+export const selectContestError = (state: RootState) => state.contest.error
